# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IRUSA Smart Event & Donor Intelligence Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{eventId}/allocations:
    get:
      operationId: listAllocations
      tags: [allocations]
      summary: List fund allocations for an event
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Allocations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Allocation"
    post:
      operationId: createAllocation
      tags: [allocations]
      summary: Create allocation
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAllocationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allocation"
  /allocations/{id}:
    patch:
      operationId: updateAllocation
      tags: [allocations]
      summary: Update allocation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAllocationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allocation"
    delete:
      operationId: deleteAllocation
      tags: [allocations]
      summary: Delete allocation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ==================== ATTENDEES ====================
  /events/{eventId}/attendees:
    get:
      operationId: listAttendees
      tags: [attendees]
      summary: List attendees for an event
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attendees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendee"
    post:
      operationId: createAttendee
      tags: [attendees]
      summary: Add attendee to event
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendeeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendee"
  /attendees/{id}:
    patch:
      operationId: updateAttendee
      tags: [attendees]
      summary: Update attendee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAttendeeBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendee"
    delete:
      operationId: deleteAttendee
      tags: [attendees]
      summary: Remove attendee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events/{eventId}/attendee-segments:
    get:
      operationId: getAttendeeSegments
      tags: [attendees]
      summary: Get attendee segmentation for an event
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Segmented attendees
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendeeSegments"

  # ==================== DONORS ====================
...
  /donations:
    get:
      operationId: listDonations
      tags: [donations]
      summary: List all donations
      parameters:
        - in: query
          name: donorId
          schema:
            type: integer
        - in: query
          name: eventId
          schema:
            type: integer
        - in: query
          name: cause
          schema:
            type: string
      responses:
        "200":
          description: Donations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Donation"
    post:
      operationId: createDonation
      tags: [donations]
      summary: Record a donation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDonationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
  /donations/{id}:
    patch:
      operationId: updateDonation
      tags: [donations]
      summary: Update donation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDonationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
    delete:
      operationId: deleteDonation
      tags: [donations]
      summary: Delete donation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ==================== FOLLOW-UP TASKS ====================
...
components:
  schemas:
...
    Allocation:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        category:
          type: string
        amount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - eventId
        - category
        - amount
        - createdAt
    CreateAllocationBody:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        notes:
          type: ["string", "null"]
      required:
        - category
        - amount
    UpdateAllocationBody:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        notes:
          type: ["string", "null"]

    # ==================== ATTENDEE ====================
    Attendee:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        attended:
          type: boolean
        donated:
          type: boolean
        donationAmount:
          type: ["number", "null"]
        volunteerInterest:
          type: boolean
        attendeeType:
          type: string
        engagementLevel:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        donorId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - eventId
        - name
        - attended
        - donated
        - volunteerInterest
        - attendeeType
        - createdAt
    CreateAttendeeBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        attended:
          type: boolean
          default: true
        donated:
          type: boolean
          default: false
        donationAmount:
          type: ["number", "null"]
        volunteerInterest:
          type: boolean
          default: false
        attendeeType:
          type: string
          default: "first-time"
        engagementLevel:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        donorId:
          type: ["integer", "null"]
      required:
        - name
    UpdateAttendeeBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        attended:
          type: boolean
        donated:
          type: boolean
        donationAmount:
          type: ["number", "null"]
        volunteerInterest:
          type: boolean
        attendeeType:
          type: string
        engagementLevel:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        donorId:
          type: ["integer", "null"]
    AttendeeSegments:
      type: object
      properties:
        attendedNotDonated:
          type: array
          items:
            $ref: "#/components/schemas/Attendee"
        attendedAndDonated:
          type: array
          items:
            $ref: "#/components/schemas/Attendee"
        firstTimeAttendees:
          type: array
          items:
            $ref: "#/components/schemas/Attendee"
        repeatAttendees:
          type: array
          items:
            $ref: "#/components/schemas/Attendee"
        volunteerInterested:
          type: array
          items:
            $ref: "#/components/schemas/Attendee"
        highValueDonorAttendees:
          type: array
          items:
            $ref: "#/components/schemas/Attendee"
      required:
        - attendedNotDonated
        - attendedAndDonated
        - firstTimeAttendees
        - repeatAttendees
        - volunteerInterested
        - highValueDonorAttendees

    # ==================== DONOR ====================
...
    Donation:
      type: object
      properties:
        id:
          type: integer
        donorId:
          type: integer
        eventId:
          type: ["integer", "null"]
        date:
          type: string
        amount:
          type: number
        cause:
          type: ["string", "null"]
        campaign:
          type: ["string", "null"]
        season:
          type: ["string", "null"]
        donationType:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - donorId
        - date
        - amount
        - donationType
        - createdAt
    CreateDonationBody:
      type: object
      properties:
        donorId:
          type: integer
        eventId:
          type: ["integer", "null"]
        date:
          type: string
        amount:
          type: number
        cause:
          type: ["string", "null"]
        campaign:
          type: ["string", "null"]
        season:
          type: ["string", "null"]
        donationType:
          type: string
          default: "one-time"
        notes:
          type: ["string", "null"]
      required:
        - donorId
        - date
        - amount
        - donationType
    UpdateDonationBody:
      type: object
      properties:
        date:
          type: string
        amount:
          type: number
        cause:
          type: ["string", "null"]
        campaign:
          type: ["string", "null"]
        season:
          type: ["string", "null"]
        donationType:
          type: string
        notes:
          type: ["string", "null"]

    # ==================== FOLLOW-UP TASK ====================
...